HW02-5 (add spring-context)

+++ src/main/java/ru/javawebinar/topjava/repository/mock/InMemoryMealRepositoryImpl.java
@@ -1,25 +1,28 @@
 package ru.javawebinar.topjava.repository.mock;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Repository;
 import ru.javawebinar.topjava.AuthorizedUser;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.model.User;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.repository.UserRepository;
+import ru.javawebinar.topjava.to.MealWithExceed;
 import ru.javawebinar.topjava.util.DateTimeUtil;
 import ru.javawebinar.topjava.util.MealsUtil;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 
-@Repository
+@Component
 public class InMemoryMealRepositoryImpl implements MealRepository {
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
     private AtomicInteger counter = new AtomicInteger(0);
@@ -27,17 +30,15 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
     @Autowired
     private UserRepository userRepository;
 
-    public InMemoryMealRepositoryImpl(UserRepository userRepository) {
+    public InMemoryMealRepositoryImpl (UserRepository userRepository){
         this.userRepository = userRepository;
+//        MealsUtil.MEALS.forEach(m -> save(m, 1));
     }
-
-//    {
-//        MealsUtil.MEALS.forEach(this::add);
-//    }
 
     @Override
     public Meal save(Meal meal, int userId) {
         User user = userRepository.get(userId);
+        if (user == null) return null;
         if (meal.isNew()) {
             meal.setId(counter.incrementAndGet());
             user.addMealId(meal.getId());
@@ -52,6 +53,7 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
     @Override
     public boolean delete(int id, int userId) {
         User user = userRepository.get(userId);
+        if (user == null) return false;
         if (user.getMealsId().contains(id)) {
             user.deleteMealId(id);
             return repository.remove(id) != null;
@@ -64,7 +66,7 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
 
         Meal meal = repository.get(id);
         User user = userRepository.get(userId);
-        if(meal != null && !user.getMealsId().contains(id)) return null;
+        if(user != null && !user.getMealsId().contains(id)) return null;
         return meal;
     }
 
@@ -76,6 +78,7 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
     @Override
     public List<Meal> getByDate(int userId, LocalDate start, LocalDate end) {
         User user = userRepository.get(userId);
+        if (user == null) return Collections.emptyList();
         return repository.values().stream().filter(m -> user.getMealsId().contains(m.getId()))
                 .filter(m -> DateTimeUtil.isBetween(m.getDateTime().toLocalDate(), start, end))
                 .collect(Collectors.toList());
